refactor: modularizes findAnonymous function
