+++ src/logic/index.js
@@ -1,2 +1,3 @@
 export * from './downloadBlob'
+export * from './findAnonymous'
 export * from './rsa'
